Move feedback from SP -> Appointment controller
This is a straight copy and paste here.
