+++ server/routes/appointments/appointmentsController.test.ts
@@ -3,10 +3,12 @@ import request from 'supertest'
 import actionPlanFactory from '../../../testutils/factories/actionPlan'
 import actionPlanAppointmentFactory from '../../../testutils/factories/actionPlanAppointment'
 import { createDraftFactory } from '../../../testutils/factories/draft'
+import deliusServiceUserFactory from '../../../testutils/factories/deliusServiceUser'
 import hmppsAuthUserFactory from '../../../testutils/factories/hmppsAuthUser'
 import initialAssessmentAppointmentFactory from '../../../testutils/factories/initialAssessmentAppointment'
 import interventionFactory from '../../../testutils/factories/intervention'
 import sentReferralFactory from '../../../testutils/factories/sentReferral'
+import serviceCategoryFactory from '../../../testutils/factories/serviceCategory'
 import supplierAssessmentFactory from '../../../testutils/factories/supplierAssessment'
 import apiConfig from '../../config'
 import AssessRisksAndNeedsService from '../../services/assessRisksAndNeedsService'
@@ -756,6 +758,817 @@ describe('POST /service-provider/action-plan/:id/sessions/:sessionNumber/edit/:d
         })
 
       expect(draftsService.deleteDraft).not.toHaveBeenCalled()
+    })
+  })
+
+  describe('GET /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/attendance', () => {
+    it('renders a page with which the Service Provider can record the Service user‘s attendance for their initial appointment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const appointment = initialAssessmentAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        durationInMinutes: 60,
+        appointmentDeliveryType: 'PHONE_CALL',
+      })
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [appointment],
+        currentAppointmentId: appointment.id,
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/attendance`)
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Add feedback')
+          expect(res.text).toContain('Appointment details')
+          expect(res.text).toContain('1 February 2021')
+          expect(res.text).toContain('1:00pm to 2:00pm')
+        })
+    })
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/attendance`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain(
+            'Attempting to add supplier assessment attendance feedback without a current appointment'
+          )
+        })
+    })
+  })
+
+  describe('POST /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/attendance', () => {
+    describe('when the Service Provider marks the Service user as having attended the initial assessment', () => {
+      it('makes a request to the interventions service to record the Service user‘s attendance and redirects to the behaviour page', async () => {
+        const referral = sentReferralFactory.assigned().build()
+        const appointment = initialAssessmentAppointmentFactory.build()
+        const updatedAppointment = initialAssessmentAppointmentFactory.build({
+          ...appointment,
+          sessionFeedback: {
+            attendance: {
+              attended: 'yes',
+              additionalAttendanceInformation: 'Alex made the session on time',
+            },
+          },
+        })
+        const supplierAssessment = supplierAssessmentFactory.build({
+          appointments: [appointment],
+          currentAppointmentId: appointment.id,
+        })
+
+        interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+        interventionsService.recordSupplierAssessmentAppointmentAttendance.mockResolvedValue(updatedAppointment)
+
+        await request(app)
+          .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/attendance`)
+          .type('form')
+          .send({
+            attended: 'yes',
+            'additional-attendance-information': 'Alex made the session on time',
+          })
+          .expect(302)
+          .expect(
+            'Location',
+            `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/behaviour`
+          )
+      })
+    })
+
+    describe('when the Service Provider marks the Service user as not having attended the initial assessment', () => {
+      it('makes a request to the interventions service to record the Service user‘s attendance and redirects to the check-your-answers page', async () => {
+        const referral = sentReferralFactory.assigned().build()
+        const appointment = initialAssessmentAppointmentFactory.build()
+        const updatedAppointment = initialAssessmentAppointmentFactory.build({
+          ...appointment,
+          sessionFeedback: {
+            attendance: {
+              attended: 'no',
+              additionalAttendanceInformation: "I haven't heard from Alex",
+            },
+          },
+        })
+        const supplierAssessment = supplierAssessmentFactory.build({
+          appointments: [appointment],
+          currentAppointmentId: appointment.id,
+        })
+        interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+        interventionsService.recordSupplierAssessmentAppointmentAttendance.mockResolvedValue(updatedAppointment)
+        await request(app)
+          .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/attendance`)
+          .type('form')
+          .send({
+            attended: 'no',
+            'additional-attendance-information': "I haven't heard from Alex",
+          })
+          .expect(302)
+          .expect(
+            'Location',
+            `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/check-your-answers`
+          )
+      })
+    })
+
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/attendance`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain(
+            'Attempting to add supplier assessment attendance feedback without a current appointment'
+          )
+        })
+    })
+  })
+
+  describe('GET /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/behaviour', () => {
+    it('renders a page with which the Service Provider can record the Service user‘s behaviour for their initial appointment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const appointment = initialAssessmentAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+      })
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [appointment],
+        currentAppointmentId: appointment.id,
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/behaviour`)
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Add behaviour feedback')
+          expect(res.text).toContain('Describe Alex&#39;s behaviour in the assessment appointment')
+          expect(res.text).toContain(
+            'If you described poor behaviour, do you want to notify the probation practitioner?'
+          )
+        })
+    })
+
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/behaviour`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain(
+            'Attempting to add initial assessment behaviour feedback without a current appointment'
+          )
+        })
+    })
+  })
+
+  describe('POST /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/behaviour', () => {
+    describe('when the Service Provider records behaviour for the initial assessment', () => {
+      it('makes a request to the interventions service to record the Service user‘s behaviour and redirects to the check-your-answers page', async () => {
+        const referral = sentReferralFactory.assigned().build()
+        const appointment = initialAssessmentAppointmentFactory.build()
+        const updatedAppointment = initialAssessmentAppointmentFactory.build({
+          ...appointment,
+          sessionFeedback: {
+            behaviour: {
+              behaviourDescription: 'They were very respectful and polite.',
+              notifyProbationPractitioner: false,
+            },
+          },
+        })
+        const supplierAssessment = supplierAssessmentFactory.build({
+          appointments: [appointment],
+          currentAppointmentId: appointment.id,
+        })
+        interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+        interventionsService.recordSupplierAssessmentAppointmentBehaviour.mockResolvedValue(updatedAppointment)
+        await request(app)
+          .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/behaviour`)
+          .type('form')
+          .send({
+            'behaviour-description': 'They were very respectful and polite.',
+            'notify-probation-practitioner': 'no',
+          })
+          .expect(302)
+          .expect(
+            'Location',
+            `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/check-your-answers`
+          )
+      })
+    })
+
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/behaviour`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain(
+            'Attempting to add initial assessment behaviour feedback without a current appointment'
+          )
+        })
+    })
+  })
+
+  describe('GET /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/check-your-answers', () => {
+    it('renders a page with which the Service Provider can view the feedback for the initial appointment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const appointment = initialAssessmentAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        sessionFeedback: {
+          attendance: {
+            attended: 'yes',
+            additionalAttendanceInformation: 'He was punctual',
+          },
+          behaviour: {
+            behaviourDescription: 'Acceptable',
+          },
+          submitted: false,
+        },
+      })
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [appointment],
+        currentAppointmentId: appointment.id,
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(
+          `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/check-your-answers`
+        )
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Confirm feedback')
+          expect(res.text).toContain('Did Alex attend the initial assessment appointment?')
+          expect(res.text).toContain('Yes, they were on time')
+          expect(res.text).toContain('Describe Alex&#39;s behaviour in the assessment appointment')
+          expect(res.text).toContain('Acceptable')
+        })
+    })
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(
+          `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/check-your-answers`
+        )
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain(
+            'Attempting to check supplier assessment feedback answers without a current appointment'
+          )
+        })
+    })
+  })
+
+  describe('POST /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/submit', () => {
+    it('submits the action plan and redirects to the confirmation page', async () => {
+      const appointment = initialAssessmentAppointmentFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [appointment],
+        currentAppointmentId: appointment.id,
+      })
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+      interventionsService.submitSupplierAssessmentAppointmentFeedback.mockResolvedValue(appointment)
+
+      await request(app)
+        .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/submit`)
+        .expect(302)
+        .expect(
+          'Location',
+          `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/confirmation`
+        )
+
+      expect(interventionsService.submitSupplierAssessmentAppointmentFeedback).toHaveBeenCalledWith(
+        'token',
+        referral.id
+      )
+    })
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const appointment = initialAssessmentAppointmentFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+      interventionsService.submitSupplierAssessmentAppointmentFeedback.mockResolvedValue(appointment)
+
+      await request(app)
+        .post(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/submit`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain('Attempting to submit supplier assessment feedback without a current appointment')
+        })
+    })
+  })
+
+  describe('GET /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback/confirmation', () => {
+    it('renders a page confirming that the supplier assessment feedback has been submitted', async () => {
+      const referral = sentReferralFactory.assigned().build()
+
+      interventionsService.getSentReferral.mockResolvedValue(sentReferralFactory.build())
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/confirmation`)
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Initial assessment added')
+        })
+    })
+  })
+
+  describe('GET /service-provider/referrals/:id/supplier-assessment/post-assessment-feedback', () => {
+    it('renders a page showing the supplier assessment feedback for the current appointment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const appointment = initialAssessmentAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        sessionFeedback: {
+          attendance: {
+            attended: 'yes',
+            additionalAttendanceInformation: 'He was punctual',
+          },
+          behaviour: {
+            behaviourDescription: 'Acceptable',
+          },
+          submitted: false,
+        },
+      })
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [appointment],
+        currentAppointmentId: appointment.id,
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback`)
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('View feedback')
+          expect(res.text).toContain('Did Alex attend the initial assessment appointment?')
+          expect(res.text).toContain('Yes, they were on time')
+          expect(res.text).toContain('Describe Alex&#39;s behaviour in the assessment appointment')
+          expect(res.text).toContain('Acceptable')
+        })
+    })
+    it('renders an error if there is no current appointment for the supplier assessment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain('Attempting to view supplier assessment feedback without a current appointment')
+        })
+    })
+  })
+
+  describe('GET /service-provider/referrals/:referralId/supplier-assessment/post-assessment-feedback/:appointmentId', () => {
+    it('renders a page showing the supplier assessment feedback for the given appointment', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const appointment = initialAssessmentAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        sessionFeedback: {
+          attendance: {
+            attended: 'no',
+            additionalAttendanceInformation: 'They missed the bus',
+          },
+          submitted: true,
+        },
+      })
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [appointment],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(
+          `/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/${appointment.id}`
+        )
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('View feedback')
+          expect(res.text).toContain('Did Alex attend the initial assessment appointment?')
+          expect(res.text).toContain('No')
+          expect(res.text).toContain('They missed the bus')
+        })
+    })
+
+    it('renders an error if an appointment is not found with that ID', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const supplierAssessment = supplierAssessmentFactory.build({
+        appointments: [],
+      })
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
+
+      await request(app)
+        .get(`/service-provider/referrals/${referral.id}/supplier-assessment/post-assessment-feedback/not-a-real-id`)
+        .expect(500)
+        .expect(res => {
+          expect(res.text).toContain('Could not find the requested appointment')
+        })
+    })
+  })
+
+  describe('GET /service-provider/action-plan/:actionPlanId/appointment/:sessionNumber/post-session-feedback/attendance', () => {
+    it('renders a page with which the Service Provider can record the Service user‘s attendance', async () => {
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const submittedActionPlan = actionPlanFactory.submitted().build({ referralId: referral.id })
+      const appointment = actionPlanAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        durationInMinutes: 60,
+        appointmentDeliveryType: 'PHONE_CALL',
+      })
+
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getActionPlanAppointment.mockResolvedValue(appointment)
+
+      await request(app)
+        .get(
+          `/service-provider/action-plan/${submittedActionPlan.id}/appointment/${appointment.sessionNumber}/post-session-feedback/attendance`
+        )
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Add attendance feedback')
+          expect(res.text).toContain('Session details')
+          expect(res.text).toContain('1 February 2021')
+          expect(res.text).toContain('1:00pm to 2:00pm')
+        })
+    })
+  })
+
+  describe('POST /service-provider/action-plan/:actionPlanId/appointment/:sessionNumber/post-session-feedback/attendance', () => {
+    describe('when the Service Provider marks the Service user as having attended the session', () => {
+      it('makes a request to the interventions service to record the Service user‘s attendance and redirects to the behaviour page', async () => {
+        const updatedAppointment = actionPlanAppointmentFactory.build({
+          sessionNumber: 1,
+          sessionFeedback: {
+            attendance: {
+              attended: 'yes',
+              additionalAttendanceInformation: 'Alex made the session on time',
+            },
+          },
+        })
+
+        const actionPlan = actionPlanFactory.build()
+
+        interventionsService.recordActionPlanAppointmentAttendance.mockResolvedValue(updatedAppointment)
+
+        await request(app)
+          .post(
+            `/service-provider/action-plan/${actionPlan.id}/appointment/${updatedAppointment.sessionNumber}/post-session-feedback/attendance`
+          )
+          .type('form')
+          .send({
+            attended: 'yes',
+            'additional-attendance-information': 'Alex made the session on time',
+          })
+          .expect(302)
+          .expect(
+            'Location',
+            `/service-provider/action-plan/${actionPlan.id}/appointment/${updatedAppointment.sessionNumber}/post-session-feedback/behaviour`
+          )
+      })
+    })
+
+    describe('when the Service Provider marks the Service user as not having attended the session', () => {
+      it('makes a request to the interventions service to record the Service user‘s attendance and redirects to the check-your-answers page', async () => {
+        const updatedAppointment = actionPlanAppointmentFactory.build({
+          sessionNumber: 1,
+          sessionFeedback: {
+            attendance: {
+              attended: 'no',
+              additionalAttendanceInformation: "I haven't heard from Alex",
+            },
+          },
+        })
+
+        const actionPlan = actionPlanFactory.build()
+
+        interventionsService.recordActionPlanAppointmentAttendance.mockResolvedValue(updatedAppointment)
+
+        await request(app)
+          .post(
+            `/service-provider/action-plan/${actionPlan.id}/appointment/${updatedAppointment.sessionNumber}/post-session-feedback/attendance`
+          )
+          .type('form')
+          .send({
+            attended: 'no',
+            'additional-attendance-information': "I haven't heard from Alex",
+          })
+          .expect(302)
+          .expect(
+            'Location',
+            `/service-provider/action-plan/${actionPlan.id}/appointment/${updatedAppointment.sessionNumber}/post-session-feedback/check-your-answers`
+          )
+      })
+    })
+  })
+
+  describe('GET /service-provider/action-plan/:actionPlanId/appointment/:sessionNumber/post-session-feedback/behaviour', () => {
+    it('renders a page with which the Service Provider can record the Service user‘s behaviour', async () => {
+      const serviceCategory = serviceCategoryFactory.build({ name: 'accommodation' })
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const submittedActionPlan = actionPlanFactory.submitted().build({ referralId: referral.id })
+      const appointment = actionPlanAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+      })
+
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getServiceCategory.mockResolvedValue(serviceCategory)
+      interventionsService.getActionPlanAppointment.mockResolvedValue(appointment)
+
+      await request(app)
+        .get(
+          `/service-provider/action-plan/${submittedActionPlan.id}/appointment/${appointment.sessionNumber}/post-session-feedback/behaviour`
+        )
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Add behaviour feedback')
+        })
+    })
+  })
+
+  describe('POST /service-provider/action-plan/:actionPlanId/appointment/:sessionNumber/post-session-feedback/behaviour', () => {
+    it('makes a request to the interventions service to record the Service user‘s behaviour and redirects to the check your answers page', async () => {
+      const updatedAppointment = actionPlanAppointmentFactory.build({
+        sessionNumber: 1,
+        sessionFeedback: {
+          behaviour: {
+            behaviourDescription: 'Alex was well-behaved',
+            notifyProbationPractitioner: false,
+          },
+        },
+      })
+
+      const actionPlan = actionPlanFactory.build()
+
+      interventionsService.recordActionPlanAppointmentBehavior.mockResolvedValue(updatedAppointment)
+
+      await request(app)
+        .post(
+          `/service-provider/action-plan/${actionPlan.id}/appointment/${updatedAppointment.sessionNumber}/post-session-feedback/behaviour`
+        )
+        .type('form')
+        .send({
+          'behaviour-description': 'Alex was well-behaved',
+          'notify-probation-practitioner': 'no',
+        })
+        .expect(302)
+        .expect(
+          'Location',
+          `/service-provider/action-plan/${actionPlan.id}/appointment/${updatedAppointment.sessionNumber}/post-session-feedback/check-your-answers`
+        )
+    })
+  })
+
+  describe('GET /service-provider/action-plan:actionPlanId/appointment/:sessionNumber/post-session-feedback/check-your-answers', () => {
+    it('renders a page with answers the user has so far selected', async () => {
+      const serviceCategory = serviceCategoryFactory.build({ name: 'accommodation' })
+      const deliusServiceUser = deliusServiceUserFactory.build()
+      const referral = sentReferralFactory.assigned().build()
+      const submittedActionPlan = actionPlanFactory.submitted().build({ referralId: referral.id })
+      const appointment = actionPlanAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        durationInMinutes: 60,
+        appointmentDeliveryType: 'PHONE_CALL',
+      })
+
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getServiceCategory.mockResolvedValue(serviceCategory)
+      interventionsService.getActionPlanAppointment.mockResolvedValue(appointment)
+
+      await request(app)
+        .get(
+          `/service-provider/action-plan/${submittedActionPlan.id}/appointment/${appointment.sessionNumber}/post-session-feedback/check-your-answers`
+        )
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('Confirm feedback')
+        })
+    })
+  })
+
+  describe('POST /service-provider/action-plan:actionPlanId/appointment/:sessionNumber/post-session-feedback/submit', () => {
+    it('marks the appointment as submitted and redirects to the confirmation page', async () => {
+      const actionPlanId = '91e7ceab-74fd-45d8-97c8-ec58844618dd'
+      const sessionNumber = 2
+
+      await request(app)
+        .post(`/service-provider/action-plan/${actionPlanId}/appointment/${sessionNumber}/post-session-feedback/submit`)
+        .expect(302)
+        .expect(
+          'Location',
+          `/service-provider/action-plan/${actionPlanId}/appointment/${sessionNumber}/post-session-feedback/confirmation`
+        )
+
+      expect(interventionsService.submitActionPlanSessionFeedback).toHaveBeenCalledWith(
+        'token',
+        actionPlanId,
+        sessionNumber
+      )
+    })
+  })
+
+  describe('GET /service-provider/action-plan/:actionPlanId/appointment/:sessionNumber/post-session-feedback/confirmation', () => {
+    describe('when final appointment attendance has been recorded', () => {
+      it('renders a page confirming that the action plan has been submitted', async () => {
+        const referral = sentReferralFactory.assigned().build()
+        const finalSessionNumber = 2
+        const submittedActionPlan = actionPlanFactory
+          .submitted()
+          .build({ referralId: referral.id, numberOfSessions: finalSessionNumber })
+
+        interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+
+        const finalAppointment = actionPlanAppointmentFactory.build({ sessionNumber: finalSessionNumber })
+
+        interventionsService.getActionPlanAppointment.mockResolvedValue(finalAppointment)
+        interventionsService.getSubsequentActionPlanAppointment.mockResolvedValue(null)
+
+        interventionsService.getSentReferral.mockResolvedValue(sentReferralFactory.build())
+
+        await request(app)
+          .get(
+            `/service-provider/action-plan/${submittedActionPlan.id}/appointment/2/post-session-feedback/confirmation`
+          )
+          .expect(200)
+          .expect(res => {
+            expect(res.text).toContain('Please submit the end of service report within 5 working days.')
+          })
+      })
+    })
+
+    describe('when any non-final session has been recorded and more sessions are scheduled', () => {
+      it('renders a page confirming that the action plan has been submitted', async () => {
+        const referral = sentReferralFactory.assigned().build()
+        const finalSessionNumber = 3
+        const submittedActionPlan = actionPlanFactory
+          .submitted()
+          .build({ referralId: referral.id, numberOfSessions: finalSessionNumber })
+
+        interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+
+        const penultimateAppointment = actionPlanAppointmentFactory.build({ sessionNumber: 2 })
+
+        const nextAppointment = actionPlanAppointmentFactory.build({
+          sessionNumber: finalSessionNumber,
+          appointmentTime: '2021-03-31T10:50:10.790Z',
+        })
+
+        interventionsService.getActionPlanAppointment.mockResolvedValue(penultimateAppointment)
+        interventionsService.getSubsequentActionPlanAppointment.mockResolvedValue(nextAppointment)
+
+        interventionsService.getSentReferral.mockResolvedValue(sentReferralFactory.build())
+
+        await request(app)
+          .get(
+            `/service-provider/action-plan/${submittedActionPlan.id}/appointment/${penultimateAppointment.sessionNumber}/post-session-feedback/confirmation`
+          )
+          .expect(200)
+          .expect(res => {
+            expect(res.text).toContain('You can now deliver the next session scheduled for 31 March 2021.')
+          })
+      })
+    })
+
+    describe('when any non-final session has been recorded but no more sessions are scheduled', () => {
+      it('renders a page confirming that the action plan has been submitted', async () => {
+        const referral = sentReferralFactory.assigned().build()
+        const finalSessionNumber = 3
+        const submittedActionPlan = actionPlanFactory
+          .submitted()
+          .build({ referralId: referral.id, numberOfSessions: finalSessionNumber })
+
+        interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+
+        const penultimateAppointment = actionPlanAppointmentFactory.build({ sessionNumber: 2 })
+
+        interventionsService.getActionPlanAppointment.mockResolvedValue(penultimateAppointment)
+        interventionsService.getSubsequentActionPlanAppointment.mockResolvedValue(null)
+
+        interventionsService.getSentReferral.mockResolvedValue(sentReferralFactory.build())
+
+        await request(app)
+          .get(
+            `/service-provider/action-plan/${submittedActionPlan.id}/appointment/${penultimateAppointment.sessionNumber}/post-session-feedback/confirmation`
+          )
+          .expect(200)
+          .expect(res => {
+            expect(res.text).toContain('The probation practitioner has been sent a copy of the session feedback form.')
+          })
+      })
+    })
+  })
+
+  describe('GET /service-provider/action-plan/:actionPlanId/appointment/:sessionNumber/post-session-feedback', () => {
+    it('renders a page displaying feedback answers', async () => {
+      const referral = sentReferralFactory.assigned().build()
+      const submittedActionPlan = actionPlanFactory.submitted().build({ referralId: referral.id, numberOfSessions: 1 })
+      const deliusServiceUser = deliusServiceUserFactory.build()
+
+      const appointmentWithSubmittedFeedback = actionPlanAppointmentFactory.build({
+        appointmentTime: '2021-02-01T13:00:00Z',
+        durationInMinutes: 60,
+        appointmentDeliveryType: 'PHONE_CALL',
+        sessionNumber: 1,
+        sessionFeedback: {
+          attendance: {
+            attended: 'yes',
+            additionalAttendanceInformation: 'They were early to the session',
+          },
+          behaviour: {
+            behaviourDescription: 'Alex was well-behaved',
+            notifyProbationPractitioner: false,
+          },
+          submitted: true,
+        },
+      })
+
+      communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
+      interventionsService.getActionPlan.mockResolvedValue(submittedActionPlan)
+      interventionsService.getSentReferral.mockResolvedValue(referral)
+      interventionsService.getActionPlanAppointment.mockResolvedValue(appointmentWithSubmittedFeedback)
+
+      await request(app)
+        .get(`/service-provider/action-plan/${submittedActionPlan.id}/appointment/1/post-session-feedback`)
+        .expect(200)
+        .expect(res => {
+          expect(res.text).toContain('View feedback')
+          expect(res.text).toContain('They were early to the session')
+          expect(res.text).toContain('Yes, they were on time')
+          expect(res.text).toContain('Alex was well-behaved')
+          expect(res.text).toContain('No')
+        })
     })
   })
 })
