+++ server/routes/serviceProviderReferrals/serviceProviderReferralsController.ts
@@ -28,17 +28,6 @@ import ActionPlanConfirmationView from '../service-provider/action-plan/confirma
 import AddActionPlanNumberOfSessionsView from '../service-provider/action-plan/number-of-sessions/actionPlanNumberOfSessionsView'
 import AddActionPlanNumberOfSessionsPresenter from '../service-provider/action-plan/number-of-sessions/actionPlanNumberOfSessionsPresenter'
 import ActionPlanNumberOfSessionsForm from '../service-provider/action-plan/number-of-sessions/actionPlanNumberOfSessionsForm'
-import AttendanceFeedbackView from '../appointments/feedback/shared/attendance/attendanceFeedbackView'
-import ActionPlanPostSessionAttendanceFeedbackPresenter from '../appointments/feedback/actionPlanSessions/attendance/actionPlanPostSessionAttendanceFeedbackPresenter'
-import AttendanceFeedbackForm from '../appointments/feedback/shared/attendance/attendanceFeedbackForm'
-import PostSessionFeedbackConfirmationPresenter from '../appointments/feedback/actionPlanSessions/confirmation/postSessionFeedbackConfirmationPresenter'
-import PostSessionFeedbackConfirmationView from '../appointments/feedback/actionPlanSessions/confirmation/postSessionFeedbackConfirmationView'
-import BehaviourFeedbackView from '../appointments/feedback/shared/behaviour/behaviourFeedbackView'
-import BehaviourFeedbackForm from '../appointments/feedback/shared/behaviour/behaviourFeedbackForm'
-import CheckFeedbackAnswersView from '../appointments/feedback/shared/checkYourAnswers/checkFeedbackAnswersView'
-import ActionPlanPostSessionFeedbackCheckAnswersPresenter from '../appointments/feedback/actionPlanSessions/checkYourAnswers/actionPlanPostSessionFeedbackCheckAnswersPresenter'
-import SubmittedFeedbackView from '../shared/appointment/feedback/submittedFeedbackView'
-import SubmittedFeedbackPresenter from '../shared/appointment/feedback/submittedFeedbackPresenter'
 import EndOfServiceReportOutcomeForm from '../service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomeForm'
 import EndOfServiceReportOutcomePresenter from '../service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomePresenter'
 import EndOfServiceReportOutcomeView from '../service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomeView'
@@ -54,23 +43,11 @@ import ServiceCategory from '../../models/serviceCategory'
 import AssessRisksAndNeedsService from '../../services/assessRisksAndNeedsService'
 import ActionPlanPresenter from '../shared/action-plan/actionPlanPresenter'
 import ActionPlanView from '../shared/action-plan/actionPlanView'
-import SupplierAssessmentDecorator from '../../decorators/supplierAssessmentDecorator'
 import ActionPlanEditConfirmationPresenter from '../service-provider/action-plan/edit/actionPlanEditConfirmationPresenter'
 import ActionPlanEditConfirmationView from '../service-provider/action-plan/edit/actionPlanEditConfirmationView'
-import InitialAssessmentAttendanceFeedbackPresenter from '../appointments/feedback/initialAssessment/attendance/initialAssessmentAttendanceFeedbackPresenter'
-import InitialAssessmentFeedbackCheckAnswersPresenter from '../appointments/feedback/initialAssessment/checkYourAnswers/initialAssessmentFeedbackCheckAnswersPresenter'
-import InitialAssessmentFeedbackConfirmationPresenter from '../appointments/feedback/initialAssessment/confirmation/initialAssessmentFeedbackConfirmationPresenter'
-import InitialAssessmentFeedbackConfirmationView from '../appointments/feedback/initialAssessment/confirmation/initialAssessmentFeedbackConfirmationView'
-import ActionPlanSessionBehaviourFeedbackPresenter from '../appointments/feedback/actionPlanSessions/behaviour/actionPlanSessionBehaviourFeedbackPresenter'
-import InitialAssessmentBehaviourFeedbackPresenter from '../appointments/feedback/initialAssessment/behaviour/initialAssessmentBehaviourFeedbackPresenter'
 import DraftsService from '../../services/draftsService'
-import {
-  ActionPlanAppointment,
-  AppointmentSchedulingDetails,
-  InitialAssessmentAppointment,
-} from '../../models/appointment'
+import { ActionPlanAppointment, AppointmentSchedulingDetails } from '../../models/appointment'
 import DeliusOfficeLocationFilter from '../../services/deliusOfficeLocationFilter'
-import AppointmentSummary from '../appointments/appointmentSummary'
 import ReferenceDataService from '../../services/referenceDataService'
 import createFormValidationErrorOrRethrow from '../../utils/interventionsFormError'
 import EndOfServiceReportPresenter from '../shared/endOfServiceReport/endOfServiceReportPresenter'
@@ -587,414 +564,6 @@ export default class ServiceProviderReferralsController {
     const view = new AddActionPlanNumberOfSessionsView(presenter)
     res.status(formError === null ? 200 : 400)
     return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async addPostSessionAttendanceFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-
-    let formError: FormValidationError | null = null
-    let userInputData: Record<string, unknown> | null = null
-
-    const data = await new AttendanceFeedbackForm(req).data()
-
-    if (req.method === 'POST') {
-      if (data.error) {
-        res.status(400)
-        formError = data.error
-        userInputData = req.body
-      } else {
-        const updatedAppointment = await this.interventionsService.recordActionPlanAppointmentAttendance(
-          accessToken,
-          actionPlanId,
-          Number(sessionNumber),
-          data.paramsForUpdate
-        )
-
-        const redirectPath =
-          updatedAppointment.sessionFeedback?.attendance?.attended === 'no' ? 'check-your-answers' : 'behaviour'
-
-        return res.redirect(
-          `/service-provider/action-plan/${actionPlanId}/appointment/${sessionNumber}/post-session-feedback/${redirectPath}`
-        )
-      }
-    }
-
-    const actionPlan = await this.interventionsService.getActionPlan(accessToken, actionPlanId)
-
-    const referral = await this.interventionsService.getSentReferral(accessToken, actionPlan.referralId)
-
-    const appointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      actionPlanId,
-      Number(sessionNumber)
-    )
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(appointment)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new ActionPlanPostSessionAttendanceFeedbackPresenter(
-      appointment,
-      serviceUser,
-      new AppointmentSummary(appointment, null, deliusOfficeLocation),
-      formError,
-      userInputData,
-      referral.id
-    )
-    const view = new AttendanceFeedbackView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async addInitialAssessmentAttendanceFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-
-    const [referral, supplierAssessment] = await Promise.all([
-      this.interventionsService.getSentReferral(accessToken, referralId),
-      this.interventionsService.getSupplierAssessment(accessToken, referralId),
-    ])
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
-    if (appointment === null) {
-      throw new Error('Attempting to add supplier assessment attendance feedback without a current appointment')
-    }
-    let formError: FormValidationError | null = null
-    let userInputData: Record<string, unknown> | null = null
-
-    const data = await new AttendanceFeedbackForm(req).data()
-
-    if (req.method === 'POST') {
-      if (data.error) {
-        res.status(400)
-        formError = data.error
-        userInputData = req.body
-      } else {
-        const updatedAppointment = await this.interventionsService.recordSupplierAssessmentAppointmentAttendance(
-          accessToken,
-          referralId,
-          data.paramsForUpdate
-        )
-        const redirectPath =
-          updatedAppointment.sessionFeedback?.attendance?.attended === 'no' ? 'check-your-answers' : 'behaviour'
-        return res.redirect(
-          `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/${redirectPath}`
-        )
-      }
-    }
-
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(appointment)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new InitialAssessmentAttendanceFeedbackPresenter(
-      appointment,
-      serviceUser,
-      new AppointmentSummary(appointment, null, deliusOfficeLocation),
-      formError,
-      userInputData,
-      referralId
-    )
-    const view = new AttendanceFeedbackView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async addInitialAssessmentBehaviourFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-
-    const [referral, supplierAssessment] = await Promise.all([
-      this.interventionsService.getSentReferral(accessToken, referralId),
-      this.interventionsService.getSupplierAssessment(accessToken, referralId),
-    ])
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
-    if (appointment === null) {
-      throw new Error('Attempting to add initial assessment behaviour feedback without a current appointment')
-    }
-    let formError: FormValidationError | null = null
-    let userInputData: Record<string, unknown> | null = null
-    if (req.method === 'POST') {
-      const data = await new BehaviourFeedbackForm(req).data()
-      if (data.error) {
-        res.status(400)
-        formError = data.error
-        userInputData = req.body
-      } else {
-        await this.interventionsService.recordSupplierAssessmentAppointmentBehaviour(
-          accessToken,
-          referralId,
-          data.paramsForUpdate
-        )
-        return res.redirect(
-          `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/check-your-answers`
-        )
-      }
-    }
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-    const presenter = new InitialAssessmentBehaviourFeedbackPresenter(
-      appointment,
-      serviceUser,
-      referralId,
-      formError,
-      userInputData
-    )
-    const view = new BehaviourFeedbackView(presenter)
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async checkInitialAssessmentFeedbackAnswers(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-
-    const [referral, supplierAssessment] = await Promise.all([
-      this.interventionsService.getSentReferral(accessToken, referralId),
-      this.interventionsService.getSupplierAssessment(accessToken, referralId),
-    ])
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
-    if (appointment === null) {
-      throw new Error('Attempting to check supplier assessment feedback answers without a current appointment')
-    }
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(appointment)
-
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-    const presenter = new InitialAssessmentFeedbackCheckAnswersPresenter(
-      appointment,
-      serviceUser,
-      referralId,
-      new AppointmentSummary(appointment, null, deliusOfficeLocation)
-    )
-    const view = new CheckFeedbackAnswersView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async submitPostAssessmentFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-    const supplierAssessment = await this.interventionsService.getSupplierAssessment(accessToken, referralId)
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
-    if (appointment === null) {
-      throw new Error('Attempting to submit supplier assessment feedback without a current appointment')
-    }
-
-    await this.interventionsService.submitSupplierAssessmentAppointmentFeedback(accessToken, referralId)
-
-    return res.redirect(
-      `/service-provider/referrals/${referralId}/supplier-assessment/post-assessment-feedback/confirmation`
-    )
-  }
-
-  async showPostAssessmentFeedbackConfirmation(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-
-    const referral = await this.interventionsService.getSentReferral(accessToken, referralId)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new InitialAssessmentFeedbackConfirmationPresenter(referralId)
-    const view = new InitialAssessmentFeedbackConfirmationView(presenter)
-
-    ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async viewPostAssessmentFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { referralId, appointmentId } = req.params
-
-    const [referral, supplierAssessment] = await Promise.all([
-      this.interventionsService.getSentReferral(accessToken, referralId),
-      this.interventionsService.getSupplierAssessment(accessToken, referralId),
-    ])
-
-    let supplierAssessmentAppointment: InitialAssessmentAppointment | null
-
-    if (appointmentId) {
-      supplierAssessmentAppointment =
-        supplierAssessment.appointments.find(appointment => appointment.id === appointmentId) || null
-
-      if (!supplierAssessmentAppointment) {
-        throw new Error('Could not find the requested appointment')
-      }
-    } else {
-      supplierAssessmentAppointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
-
-      if (supplierAssessmentAppointment === null) {
-        throw new Error('Attempting to view supplier assessment feedback without a current appointment')
-      }
-    }
-
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(
-      supplierAssessmentAppointment
-    )
-
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new SubmittedFeedbackPresenter(
-      supplierAssessmentAppointment,
-      new AppointmentSummary(supplierAssessmentAppointment, null, deliusOfficeLocation),
-      serviceUser,
-      'service-provider',
-      referralId
-    )
-    const view = new SubmittedFeedbackView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async addPostSessionBehaviourFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-
-    let formError: FormValidationError | null = null
-    let userInputData: Record<string, unknown> | null = null
-
-    if (req.method === 'POST') {
-      const data = await new BehaviourFeedbackForm(req).data()
-
-      if (data.error) {
-        res.status(400)
-        formError = data.error
-        userInputData = req.body
-      } else {
-        await this.interventionsService.recordActionPlanAppointmentBehavior(
-          accessToken,
-          actionPlanId,
-          Number(sessionNumber),
-          data.paramsForUpdate
-        )
-
-        return res.redirect(
-          `/service-provider/action-plan/${actionPlanId}/appointment/${sessionNumber}/post-session-feedback/check-your-answers`
-        )
-      }
-    }
-
-    const actionPlan = await this.interventionsService.getActionPlan(accessToken, actionPlanId)
-    const referral = await this.interventionsService.getSentReferral(accessToken, actionPlan.referralId)
-
-    const appointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      actionPlanId,
-      Number(sessionNumber)
-    )
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new ActionPlanSessionBehaviourFeedbackPresenter(
-      appointment,
-      serviceUser,
-      actionPlanId,
-      formError,
-      userInputData
-    )
-    const view = new BehaviourFeedbackView(presenter)
-
-    res.status(formError === null ? 200 : 400)
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async checkPostSessionFeedbackAnswers(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-
-    const actionPlan = await this.interventionsService.getActionPlan(accessToken, actionPlanId)
-    const referral = await this.interventionsService.getSentReferral(accessToken, actionPlan.referralId)
-
-    const currentAppointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      actionPlanId,
-      Number(sessionNumber)
-    )
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(currentAppointment)
-
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new ActionPlanPostSessionFeedbackCheckAnswersPresenter(
-      currentAppointment,
-      serviceUser,
-      actionPlanId,
-      new AppointmentSummary(currentAppointment, null, deliusOfficeLocation)
-    )
-    const view = new CheckFeedbackAnswersView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async submitPostSessionFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-
-    await this.interventionsService.submitActionPlanSessionFeedback(accessToken, actionPlanId, Number(sessionNumber))
-
-    return res.redirect(
-      `/service-provider/action-plan/${actionPlanId}/appointment/${sessionNumber}/post-session-feedback/confirmation`
-    )
-  }
-
-  async viewSubmittedPostSessionFeedback(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-
-    const actionPlan = await this.interventionsService.getActionPlan(accessToken, actionPlanId)
-    const referral = await this.interventionsService.getSentReferral(accessToken, actionPlan.referralId)
-
-    const currentAppointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      actionPlanId,
-      Number(sessionNumber)
-    )
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(currentAppointment)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new SubmittedFeedbackPresenter(
-      currentAppointment,
-      new AppointmentSummary(currentAppointment, null, deliusOfficeLocation),
-      serviceUser,
-      'service-provider',
-      referral.id
-    )
-    const view = new SubmittedFeedbackView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async showPostSessionFeedbackConfirmation(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const { actionPlanId, sessionNumber } = req.params
-    const actionPlan = await this.interventionsService.getActionPlan(accessToken, actionPlanId)
-
-    const currentAppointment = await this.interventionsService.getActionPlanAppointment(
-      accessToken,
-      actionPlanId,
-      Number(sessionNumber)
-    )
-
-    const nextAppointmentOrNull = await this.interventionsService.getSubsequentActionPlanAppointment(
-      accessToken,
-      actionPlan,
-      currentAppointment
-    )
-
-    const referral = await this.interventionsService.getSentReferral(accessToken, actionPlan.referralId)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new PostSessionFeedbackConfirmationPresenter(
-      actionPlan,
-      currentAppointment,
-      nextAppointmentOrNull
-    )
-    const view = new PostSessionFeedbackConfirmationView(presenter)
-
-    ControllerUtils.renderWithLayout(res, view, serviceUser)
   }
 
   async createDraftEndOfServiceReport(req: Request, res: Response): Promise<void> {
